+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 * Add api for `orders`, `profile`, `liches` and `rivens`.
+* Add `experimental` API, including `auctions`.
 * Add `exceptions` wrapper.
 * Add websocket support to `Session`.
 * Add shpinx documentation page(TODO).